Improve tests for FibonacciSequenceGenerator
Firstly, the tests file for sequence generators was moved to the
location of sequence generators in the lib instead of the dedicated
tests directory. Eventually, the dedicated tests module will be
eliminated completely. The required data for the tests was also
moved similarly. A name correction was also done during this.

The tests were re-written to use more pytest-specific features
like fixtures and parametrize to eliminate the unnecessary class.
More tests were added to include edge cases and make safer the
implementation of FibonacciSequenceGenerator.

Desired checked exceptions were added to the class under
consideration.